+++ DeBroglie.Console/MagicaVoxelLoader.cs
@@ -1,5 +1,4 @@
 ﻿using DeBroglie.MagicaVoxel;
-using DeBroglie.Models;
 using DeBroglie.Topo;
 using System.IO;
 
@@ -27,22 +26,6 @@ namespace DeBroglie.Console
         public Tile Parse(string s)
         {
             return new Tile(byte.Parse(s));
-        }
-    }
-
-    public class MagicaVoxelSaver : ISampleSetSaver
-    { 
-        public void Save(TileModel model, TilePropagator tilePropagator, string filename, DeBroglieConfig config, object template)
-        {
-            var vox = template as Vox;
-            var array = tilePropagator.ToValueArray<byte>();
-            VoxUtils.Save(vox, array);
-
-            using (var stream = new FileStream(filename, FileMode.Create))
-            {
-                var br = new BinaryWriter(stream);
-                VoxSerializer.Write(br, vox);
-            }
         }
     }
 }